refactor(frontend): rename component to load Ethereum balances (#3969)
# Motivation

It makes sense to rename the component `LoaderBalances` to
`LoaderEthBalances` and move it to the `eth` folder (it was previously
in `icp-eth`).
