+++ src/frontend/src/eth/components/loaders/LoaderEthBalances.svelte
