+++ src/frontend/src/lib/components/loaders/Loaders.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import LoaderBalances from '$icp-eth/components/core/LoaderBalances.svelte';
+	import LoaderEthBalances from '$eth/components/loaders/LoaderEthBalances.svelte';
 	import ExchangeWorker from '$lib/components/exchange/ExchangeWorker.svelte';
 	import AddressGuard from '$lib/components/guard/AddressGuard.svelte';
 	import Loader from '$lib/components/loaders/Loader.svelte';
@@ -10,7 +10,7 @@
 
 <AddressGuard>
 	<Loader>
-		<LoaderBalances>
+		<LoaderEthBalances>
 			<LoaderWallets>
 				<ExchangeWorker>
 					<LoaderMetamask
@@ -20,6 +20,6 @@
 					</LoaderMetamask>
 				</ExchangeWorker>
 			</LoaderWallets>
-		</LoaderBalances>
+		</LoaderEthBalances>
 	</Loader>
 </AddressGuard>
